+++ config/urls.py
@@ -33,6 +33,7 @@ urlpatterns = [
     url(r"^userprofile/", include("userprofile.urls")),
     # The rendered Sphinx-based manual.
     url(r"^manual/", include("docs.urls")),
+    url(r"^su/", include("django_su.urls")),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
 
 if settings.DEBUG: